Move menu into it's own component
